+++ Source/Coco.Server/Coco.Entities/Dtos/Auth/RoleDto.cs
@@ -4,7 +4,7 @@ namespace Coco.Entities.Dtos.Auth
 {
     public class RoleDto
     {
-        public byte Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public DateTime CreatedDate { get; set; }
