+++ Source/Coco.Server/Coco.Management/Models/RoleViewModel.cs
@@ -4,7 +4,7 @@ namespace Coco.Management.Models
 {
     public class RoleViewModel
     {
-        public byte Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public DateTime CreatedDate { get; set; }
